credit_approval_system/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
celery
redis
python-dotenv
pandas
openpyxl
python-dateutil